i32 -> usize on many occasions

+++ src/consts.rs
@@ -1,6 +1,6 @@
 
-pub const RESTART_MULTIPLIER : i32 = 512;
-pub const DECAY_INTERVAL : i32 = 128;
+pub const RESTART_MULTIPLIER : usize = 512;
+pub const DECAY_INTERVAL : usize = 128;
 pub const PHASE_THRESHOLD : i32 = 32;
 pub const CLAUSE_LIMIT : i32 = 256;
-pub const STEP_INITIAL : i32 = 1024;
+pub const STEP_INITIAL : usize = 1024;
